Cambios
Se agrega el ejemplo 2 de Pilas y Cola

+++ Ejemplo1_CG/Assets/Scripts/Punto2D.cs
@@ -9,20 +9,20 @@ namespace Package2D
 
     public class Punto2D
     {
-        public double x;
-        public double y;
+        public float x;
+        public float y;
 
 
         public Punto2D()
         {
         }
-        public Punto2D(double x, double y)
+        public Punto2D(float x, float y)
         {
             this.x = x;
             this.y = y;
         }
-        public double X { get => x; set => x = value; }
-        public double Y { get => y; set => y = value; }
+        public float X { get => x; set => x = value; }
+        public float Y { get => y; set => y = value; }
         public double Distancia(Punto2D otroPunto)
         {
             return Math.Sqrt(Math.Pow(otroPunto.X - this.X, 2) + Math.Pow(otroPunto.Y - this.Y, 2));
